Fix glitched Bottle Refill Graphics in Cap Shop
This fix changes the standing item graphics for bottle refills. Since it
doesn't make sense for those items to be anywhere other than shops, and
for them to appear the same as a filled bottle, this seems like a safe
change to make. And since the standing item graphics now point to the
bottle refill sprite, its unnecessary to use the preloaded potion refill
sprite in the shop sprite drawing routing, so that was removed as well.